chore(pages): Refactor the pages structure to add a comprehensive logic layer for the respective pages interfaces & UserAreaInterface components structure
