+++ app/UserArea/page.tsx
@@ -1,44 +1,12 @@
-"use client";
-
 import { Footer } from "@/components/Footer";
 import { Header } from "@/components/Header";
-import { AreaBackground } from "@/components/HomePageInterface/components/UserOfflineHomePage/components/AreaBackground";
-import { UserAreaInterface } from "@/components/UserAreaInterface";
-import { useAuth } from "@/context/AuthContext/AuthContext";
-import { useGlobalQuota } from "@/context/GlobalQuotaContext/GlobalQuotaContext";
-import { useUser } from "@/context/UserContext/UserContext";
-import { redirect } from "next/navigation";
-import { useMemo } from "react";
+import { UserAreaContent } from "@/components/UserAreaInterface/components/UserAreaContent";
 
 const UserArea = () => {
-  const { user, logout } = useAuth();
-  const { data, loading } = useUser();
-  const { quota } = useGlobalQuota();
-
-  if (!user) {
-    redirect("/");
-  }
-
-  const history = useMemo(
-    () => data?.historyPreview ?? [],
-    [data?.historyPreview],
-  );
-
-  const spins = quota;
-
   return (
     <>
       <Header />
-      <AreaBackground>
-        <UserAreaInterface
-          historyPreview={history}
-          loading={loading}
-          spins={spins}
-          data={data}
-          user={user}
-          logout={logout}
-        />
-      </AreaBackground>
+      <UserAreaContent />
       <Footer />
     </>
   );
