+++ app/UserSubscriptions/page.tsx
@@ -1,16 +1,15 @@
-"use client";
 import { Footer } from "@/components/Footer";
 import { Header } from "@/components/Header";
-import { UserSubscriptionsInterface } from "@/components/UserSubscriptionsInterface";
-import { useSearchParams } from "next/navigation";
+import UserSubscriptionsContent from "@/components/UserSubscriptionsInterface/components/UserSubscriptionsContent/UserSubscriptionsContent";
+import { Suspense } from "react";
 
 const UserSubscriptions = () => {
-  const searchParams = useSearchParams();
-  const planId = searchParams.get("plan");
   return (
     <>
       <Header />
-      <UserSubscriptionsInterface planId={planId} />
+      <Suspense fallback={null}>
+        <UserSubscriptionsContent />
+      </Suspense>
       <Footer />
     </>
   );
