+++ app/page.tsx
@@ -7,13 +7,10 @@ export default function Home() {
     <>
       <Header />
       <div className="bg-gradient-to-br from-amber-500 to-[#84e9e4]">
-        <div>
-          <main className="flex flex-col justify-between  min-h-screen relative z-10 ">
-            <HomePageInterface />
-          </main>
-        </div>
+        <main className="flex flex-col justify-between  min-h-screen relative z-10 ">
+          <HomePageInterface />
+        </main>
       </div>
-
       <Footer />
     </>
   );
