+++ components/ForTenantsInterface/components/ForTenantsContent/index.ts
@@ -1,0 +1,1 @@
+export { default as ForTenantsContent } from "./ForTenantsContent";
