+++ components/Slots/components/TenantSlotsContent/index.ts
@@ -1,0 +1,1 @@
+export { default as TenantSlotsContent } from "./TenantSlotsContent";
