+++ components/TenantAreaInterface/components/TenantAreaContent/index.ts
@@ -1,0 +1,1 @@
+export { default as TenantAreaContent } from "./TenantAreaContent";
