+++ components/UserAreaInterface/components/DailyQuota/components/DailyOffersHistory/DailyOffersHistory.tsx
