+++ components/UserAreaInterface/components/DailyQuota/components/DailyOffersHistory/components/HistoryProductItem/HistoryProductItem.tsx
