+++ components/UserAreaInterface/components/DailyQuota/components/DailyOffersHistory/components/HistoryProductItem/index.ts
