+++ components/UserAreaInterface/components/DailyQuota/components/DailyOffersHistory/components/ProductHistoryCard/index.ts
