+++ components/UserAreaInterface/components/DailyQuota/components/DailyOffersHistory/index.ts
