+++ components/UserAreaInterface/components/DailyQuota/components/OutOfSpinsInterface/index.ts
