+++ components/UserAreaInterface/components/DailyQuota/components/RemainingDisplay/index.ts
