+++ components/UserAreaInterface/components/Rewards/index.ts
