+++ components/UserAreaInterface/components/UserAreaContent/index.ts
@@ -1,0 +1,1 @@
+export { default as UserAreaContent } from "./UserAreaContent";
