+++ components/UserAreaInterface/components/UserStats/components/AccountRewardHistory/index.ts
