+++ components/UserAreaInterface/components/UserStats/components/TrophiesAcquired/index.ts
