+++ components/UserSubscriptionsInterface/components/UserSubscriptionsContent/index.ts
@@ -1,0 +1,1 @@
+export { default as UserSubscriptionContent } from "./UserSubscriptionsContent";
